+++ LibraryComponents/Classes/AndesThumbnail/Hierarchy/AndesThumbnailHierarchyProtocol.swift
@@ -1,0 +1,15 @@
+//
+//  AndesThumbnailHierarchyProtocol.swift
+//  AndesUI
+//
+//  Created by Alejo Echeguia on 24/06/2020.
+//
+
+import Foundation
+
+internal protocol AndesThumbnailHierarchyProtocol {
+	var backgroundColor: UIColor { get }
+	var iconColor: UIColor? { get }
+	var borderColor: UIColor? { get }
+	var overlayColor: UIColor? { get }
+}