+++ LibraryComponents/Classes/AndesThumbnail/Type/AndesThumbnailTypeProtocol.swift
@@ -1,0 +1,15 @@
+//
+//  AndesThumbnailTypeProtocol.swift
+//  AndesUI
+//
+//  Created by Alejo Echeguia on 30/06/2020.
+//
+
+import Foundation
+
+internal protocol AndesThumbnailTypeProtocol {
+	var accentColor: UIColor? { get }
+	var borderColor: UIColor? { get }
+	var circular: Bool { get }
+
+}