+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v3.6.0
 ### 🚀 Features
+- New component: AndesThumbnail | Authors: [@alejoecheguia](https://github.com/alejoecheguia)
 - Andes Message: Now allows to have a `Link Action` in the place of primary button, it has the same function as a button with a new layout defined by Andes's UX. | Authors: [@lelalesi](https://github.com/lelalesi)
 
 ## v3.5.0